@@ -1,8 +1,8 @@
 {
   "name": "postbot",
-  "version": "0.1.0-beta",
+  "version": "0.1.0-beta_dev",
   "description": "A TS chess engine",
-  "type": "commonjs",
+  "type": "module",
   "main": "compiled/main.js",
   "scripts": {
     "start": "node compiled/main.js",
@@ -13,29 +13,30 @@
   "pkg": {
     "scripts": "compiled/chessWorker.js",
     "targets": [
-      "latest-macos-x64",
-      "latest-linux-x64",
-      "latest-win-x64"
+      "latest-macos-arm64",
+      "latest-linux-arm64",
+      "latest-win-arm64"
     ],
     "outputPath": "build"
   },
   "author": "postnerd",
   "license": "ISC",
   "devDependencies": {
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-typescript": "^8.3.3",
-    "@types/jest": "^28.1.4",
-    "@types/node": "^18.0.3",
-    "@typescript-eslint/eslint-plugin": "^5.30.5",
-    "@typescript-eslint/parser": "^5.30.5",
-    "eslint": "^8.19.0",
-    "eslint-plugin-jest": "^26.5.3",
-    "jest": "^28.1.2",
-    "pkg": "^5.7.0",
-    "rollup": "^2.76.0",
-    "ts-jest": "^28.0.5",
-    "tslib": "^2.4.0",
-    "typescript": "^4.7.4"
+    "@rollup/plugin-json": "^6.0.1",
+    "@rollup/plugin-typescript": "^11.1.5",
+    "@types/jest": "^29.5.6",
+    "@types/node": "^20.8.7",
+    "@typescript-eslint/eslint-plugin": "^6.8.0",
+    "@typescript-eslint/parser": "^6.8.0",
+    "eslint": "^8.51.0",
+    "eslint-plugin-jest": "^27.4.2",
+    "jest": "^29.7.0",
+    "npm-check-updates": "^16.14.6",
+    "pkg": "^5.8.1",
+    "rollup": "^4.1.4",
+    "ts-jest": "^29.1.1",
+    "tslib": "^2.6.2",
+    "typescript": "^5.2.2"
   },
   "dependencies": {
     "random-js": "^2.1.0"
