@@ -1,6 +1,14 @@
-/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
-
-module.exports = {
-	preset: "ts-jest",
-	testEnvironment: "node",
+export default {
+	preset: "ts-jest/presets/default-esm",
+	moduleNameMapper: {
+		"^(\\.{1,2}/.*)\\.js$": "$1",
+	},
+	transform: {
+		"^.+\\.tsx?$": [
+			"ts-jest",
+			{
+				useESM: true,
+			},
+		],
+	},
 };
