@@ -1,4 +1,4 @@
-import Board from "../../src/board";
+import Board from "../../src/board.js";
 
 const checks = [
 	{ fen: "8/8/8/3k1b2/8/3KN3/8/8 w - - 0 1", result: true },
