@@ -1,4 +1,4 @@
-import Board from "../../src/board";
+import Board from "../../src/board.js";
 
 const checks = [
 	{ fen: "3k4/8/4n3/8/3K4/8/8/8 w - - 0 1", result: true },
